pyproject.toml
[tool.poetry]
name = "wishlist"
version = "0.1.0"
description = ""
authors = ["Thais Bighetti"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
django = "4.2.15"
djangorestframework = "^3.15.2"
model-bakery = "^1.19.4"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
django-rest-authtoken = "^2.1.4"
coverage = "^7.6.1"
pip = "^24.2"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
